requirements.txt

opencv-python
numpy
tensorflow~=2.16
mss
Pillow
pyautogui
matplotlib
